Create module for shared configuration
